Move StreamingComponent interface to its own file.
