+++ detection/api/CMakeLists.txt
@@ -37,6 +37,7 @@ find_package(OpenCV 3.3 EXACT REQUIRED PATHS /opt/opencv-3.3.0
 include_directories(include)
 set(SOURCE_FILES
         include/MPFDetectionComponent.h
+        include/MPFStreamingDetectionComponent.h
 
         include/detectionComponentUtils.h
         src/detectionComponentUtils.cpp
